cloud-run-single/apps/chat/requirements.txt
google-cloud-aiplatform>=1.38.1  # Check for the latest version
protobuf>=3.19.5                 # Often needed by google-cloud libraries
vertexai>=0.1.0                  # New unified Vertex AI SDK (includes language models)
fastapi==0.115.12
uvicorn[standard]==0.34.2
google-cloud-logging==3.12.1
